Rename Concurrent.hpp to Queue.hpp
The file now only contains Queue and CopyQueue (the mutex classes were
removed in the previous commit). Rename to reflect its actual contents.
