+++ components/devices/src/ConfigUpdate.cpp
@@ -2,7 +2,7 @@
 #include "config/ConfigStateStore.hpp"
 #include "NvsStore.hpp"
 #include "mqtt/MqttRoot.hpp"
-#include "Concurrent.hpp"
+#include "Queue.hpp"
 #include "mqtt/MqttDriver.hpp"
 #include <ConfigUpdate.hpp>
 #include <Log.hpp>
