+++ components/devices/src/SyncPublisher.cpp
@@ -1,5 +1,5 @@
-#include "Concurrent.hpp"
 #include "KernelStatus.hpp"
+#include "Queue.hpp"
 #include "Task.hpp"
 #include "config/ConfigState.hpp"
 #include "functions/Function.hpp"
