+++ components/kernel/src/Console.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <Concurrent.hpp>
 #include <Log.hpp>
+#include <Queue.hpp>
 #include <utility>
 
 namespace cornucopia::ugly_duckling::kernel {
