+++ components/kernel/src/drivers/LedDriver.hpp
@@ -4,8 +4,8 @@
 #include <chrono>
 #include <vector>
 
-#include <Concurrent.hpp>
 #include <Pin.hpp>
+#include <Queue.hpp>
 #include <Task.hpp>
 
 using namespace std::chrono;
