+++ components/kernel/src/drivers/WiFiDriver.hpp
@@ -15,8 +15,8 @@
 
 #include "WifiApRecord.hpp"
 #include <ArduinoJson.h>
-#include <Concurrent.hpp>
 #include <Overloaded.hpp>
+#include <Queue.hpp>
 #include <State.hpp>
 #include <StateManager.hpp>
 #include <Task.hpp>
