+++ components/kernel/src/mqtt/PendingMessages.hpp
@@ -4,7 +4,7 @@
 #include <mutex>
 #include <unordered_map>
 
-#include <Concurrent.hpp>
+#include <Queue.hpp>
 
 namespace cornucopia::ugly_duckling::kernel::mqtt {
