+++ components/peripherals/src/peripherals/door/Door.hpp
@@ -8,9 +8,9 @@
 #include <utility>
 #include <variant>
 
-#include <Concurrent.hpp>
 #include <Named.hpp>
 #include <Overloaded.hpp>
+#include <Queue.hpp>
 #include <Task.hpp>
 #include <Telemetry.hpp>
 #include <Watchdog.hpp>
